Phase 1-5 completion: infrastructure, NIF pipeline, Blender integration, ObjectWindow refactor, error handling
- Build fixed after git checkout damage (240 linker errors resolved)
- Phase 1: ConfigPaths centralized, temp files cleaned, .gitignore updated, STATUS.md synced
- Phase 2: All NifPyFileWrapper methods implemented (was stubs)
- Phase 3: All BlenderLauncher methods implemented, minimum version 4.4+
- Phase 4: ObjectWindow refactored with shared getModelPathForRecord helper
- Phase 5: Error handling fixes across 7 sites, Blender exit code checking added
- 20/20 tests passing

+++ libs/files/esm/variant.cpp
@@ -131,22 +131,22 @@ void Variant::load(ESMReader& esm, Format format, const QString& editorId)
 
         if (editorId.size() > 0)
         {
-            if (editorId.front() == "b")
+            if (editorId.front() == 'b')
             {
                 setType(Var_Bool);
                 setBool(esm.readType<quint32>());
             }
-            else if (editorId.front() == "i")
+            else if (editorId.front() == 'i')
             {
                 setType(Var_Int);
-                setFloat(esm.readType<quint32>());
+                setInt(esm.readType<quint32>());
             }
-            else if (editorId.front() == "f")
+            else if (editorId.front() == 'f')
             {
                 setType(Var_Float);
                 setFloat(esm.readType<float>());
             }
-            else if (editorId.front() == "s")
+            else if (editorId.front() == 's')
             {
                 if (esm.localised())
                 {
